Split up test files
In an attempt to have a better overview of the tests
